NumericalSuite

Documentation for NumericalSuite.

NumericalSuite.Euler

NumericalSuite.MyODEProblem

NumericalSuite.MyODESolution

NumericalSuite.RK2

NumericalSuite.solve

NumericalSuite.Euler — Type

Euler <: ODEAlgorithm

The classical Euler method for solving ordinary differential equations. It is a first-order method that discretizes the time derivative as a forward difference as follows:

uₙ₊₁ = uₙ + Δt * f(uₙ, p, tₙ)

NumericalSuite.MyODEProblem — Type

MyODEProblem{uType, tType, pType, F}

Defines an Ordinary Differential Equation Problem.

To define an ODE problem, you need to provide the function f, the initial condition u0 and the time span tspan. The ODE problem is then of the form:

\[u'(t) = f(u(t), p, t), u(t_0) = u_0,\]

where t_0 is the initial time and t_1 is the final time. The optional parameter p is a vector of parameters that can be passed to the function f.

Fields

- `f`: Function that defines the ODE problem
- `u0`: Initial condition
- `tspan`: Time span
- `p`: Vector of parameters

Example Problem

using NumericalSuite

f(u, p, t) = [u[2], t - u[1]]

u0 = [1.0, 2.0]
tspan = (0.0, 1.0)

prob = MyODEProblem(f, u0, tspan)

NumericalSuite.MyODESolution — Type

MyODESolution{uType, tType, P, A, IType}

Representation of an Ordinary Differential Equation solution defined by a MyODEProblem.

The solution is represented by a vector of times t and a matrix of solutions u. Additionally, an interpolator interp is provided to interpolate the solution at any point in the time span, with the same accuracy as the original solver.

Fields

- `u`: Matrix of solutions
- `t`: Vector of times
- `prob`: MyODEProblem that defines the solution
- `alg`: ODEAlgorithm used to solve the problem
- `interp`: Interpolator used to interpolate the solution

Example of use

using NumericalSuite

f(u, p, t) = [u[2], t - u[1]]

u0 = [1.0, 2.0]
tspan = (0.0, 1.0)

prob = MyODEProblem(f, u0, tspan)
sol = solve(prob, Euler(), Δt = 0.1)

sol(0.5)

u = sol.u
t = sol.t

NumericalSuite.RK2 — Type

RK2 <: ODEAlgorithm

The second-order Runge-Kutta method for solving ordinary differential equations. It is a second-order method that discretizes the time derivative as a midpoint method as follows:

k₁ = f(uₙ, p, tₙ)

k₂ = f(uₙ + Δt * k₁, p, tₙ + Δt)

uₙ₊₁ = uₙ + Δt * (k₁ + k₂) / 2

NumericalSuite.solve — Method

solve(prob::MyODEProblem, alg::ODEAlgorithm; Δt = error("Δt not specified"))

Solve the MyODEProblem prob using the ODEAlgorithm alg and the time step Δt.

Example of use

using NumericalSuite

f(u, p, t) = [u[2], t - u[1]]

u0 = [1.0, 2.0]
tspan = (0.0, 1.0)

prob = MyODEProblem(f, u0, tspan)
sol = solve(prob, Euler(), Δt = 0.1)